♻️ move event_bus specificities in dedicated module
